Added a downloader for the implant.
The idea is to upload the compiled implant to a remote host. The downloader would be compiled and bound to an innocent executable. When the downloader executes it's going to run shellcode to download the implant to a location you specify in advance, after it has been downloaded the shellcode will execute the implant. Lastly a registry entry will be made that points to the path to which the implant was downloaded for persistence.

Shellcode for dowloading and ececuting a file can be generated using the metasploit framework's msfvenom which allows you to encode the shellcode as well to avoid anti virus detection.

Please note that this is a work in progress.